Move pkg/ext-proc -> cmd/ext-proc (#368)
* Move pkg/ext-proc -> cmd/ext-proc

* Rework Dockerfile

- Cache dependencies
- Upload only the files needed
